Modificaciones de metodos, pagar, guardarTarjetas y obtenerVuelosFuturos y historial de vuelos

+++ backend/src/main/java/hn/unah/backend/dtos/TarjetaDto.java
@@ -16,7 +16,7 @@ public class TarjetaDto {
 
     private int codigoTarjeta;
 
-    private Long numero;
+    private int numero;
 
     private String nombre;
